added bag_processing package
