+++ tools/bag_processing/scripts/check_bags_for_gaps.py
