+++ tools/bag_processing/scripts/convert_all_bayer_bags_to_grayscale.py
