+++ tools/bag_processing/scripts/trim_bag.py
